runs DI automatically new paradigms of step functions
