+++ cgk_calcium_tools/filehandler.py
@@ -6,8 +6,7 @@ import warnings
 import numpy as np
 import isx
 import json
-from typing import Union, Tuple, Iterable
-from .processing import fix_frames
+from typing import Union, Tuple
 import shutil
 from .files_io import (
     write_log_file,
@@ -24,7 +23,7 @@ from .isx_aux_functions import (
     create_empty_cellset,
     create_empty_events,
 )
-
+from .pipeline_functions import f_register, f_message
 
 def ifstr2list(x) -> list:
     if isinstance(x, list):
@@ -92,7 +91,7 @@ class isx_files_handler:
         outputsfolders: Union[str, list] = ".",
         data_subfolders: Union[str, list] = ".",
         files_patterns: Union[str, list] = "**/*.isxd",
-        processing_steps: list = ["DI", "PP", "TR", "BP", "MC", "DFF", "PM"],
+        processing_steps: list = ["PP", "TR", "BP", "MC", "DFF", "PM"],
         single_file_match: bool = False,
         recording_labels: Union[list, None] = None,
         check_new_inputs: bool = True,
@@ -104,7 +103,6 @@ class isx_files_handler:
     ):
 
         self.processing_steps = processing_steps
-        self.deinterleave_output_files = []
 
         # Check if the parameters file exists at the given path, open and load it.
         assert os.path.exists(parameters_path), "parameters file does not exist"
@@ -381,27 +379,18 @@ class isx_files_handler:
         assert (
             operation in self.processing_steps
         ), f"operation must be in the list {self.processing_steps}, got: {operation}"
-
-        # Copy the list of processing steps to a local variable
-        local_processing_steps = self.processing_steps
-
-        # Remove DI from list, if its first
-        if local_processing_steps[0] == "DI":
-            local_processing_steps.remove("DI")
-
-        opi = local_processing_steps.index(operation)
+        opi = self.processing_steps.index(operation)
 
         outputs = []
         inputs = []
 
-        suffix_out = "-" + "-".join(local_processing_steps[: opi + 1]) + ".isxd"
+        suffix_out = "-" + "-".join(self.processing_steps[: opi + 1]) + ".isxd"
 
         if opi == 0:
             suffix_in = None
-            # inputs = self.p_rec_paths
-            inputs = self.deinterleave_output_files
+            inputs = self.p_rec_paths
         else:
-            suffix_in = "-" + "-".join(local_processing_steps[:opi]) + ".isxd"
+            suffix_in = "-" + "-".join(self.processing_steps[:opi]) + ".isxd"
 
         for ofolder, file in zip(self.p_outputsfolders, self.p_rec_paths):
             outputs.append(str(Path(ofolder, Path(file).stem + suffix_out)))
@@ -554,9 +543,7 @@ class isx_files_handler:
         None
 
         """
-        if op == "DI":
-            paths = self.deinterleave_output_files
-        elif op == "DFF":
+        if op == "DFF":
             paths = self.get_results_filenames("dff", op=None)
         elif op == "PM":
             paths = self.get_results_filenames("maxdff", op=None)
@@ -600,12 +587,7 @@ class isx_files_handler:
         """
 
         if pairlist is None:
-            if op == "MC":
-                translation_files = self.get_results_filenames(
-                    "translations.csv", op=op
-                )
-                crop_rect_files = self.get_results_filenames("crop_rect.csv", op=op)
-            elif op == "DFF":
+            if op == "DFF":
                 input = self.get_filenames(op=None)
                 output = self.get_results_filenames("dff", op=None)
                 amount_of_files = len(input)
@@ -617,20 +599,12 @@ class isx_files_handler:
                 pairlist = zip(input, output)
             if op in ["PP", "TR", "BP", "MC"]:
                 amount_of_files, pairlist = self.get_pair_filenames(op)
-        elif op == "MC":
-            translation_files = [
-                os.path.splitext(p[1])[0] + "-translations.csv" for p in pairlist
-            ]
-            crop_rect_files = [
-                os.path.splitext(p[1])[0] + "-crop_rect.csv" for p in pairlist
-            ]
 
         if overwrite:
             for _, output in pairlist:
                 remove_file_and_json(output)
 
         steps_list = {
-            "DI": "de-interleave",
             "PP": "preprocessing",
             "BP": "spatial_filter",
             "MC": "motion_correct",
@@ -643,53 +617,28 @@ class isx_files_handler:
         assert len(operation) == 1, f"Step {op} not starts with {steps_list}."
         operation = operation[0]
 
-        if op != "DI":
-            parameters = self.default_parameters[operation].copy()
-            for key, value in kws.items():
-                assert key in parameters, f"The parameter: {key} does not exist"
-                parameters[key] = value
+        parameters = self.default_parameters[operation].copy()
+        for key, value in kws.items():
+            assert key in parameters, f"The parameter: {key} does not exist"
+            parameters[key] = value
 
-        # TASK: No need for this step at this point - Can you confirm Fernando?
-        '''
-        if self.processing_steps.index(op) == 0:  # if first step
-            # check if all exist
-            for finput, _ in self.get_pair_filenames(op):
-                assert os.path.exists(
-                    finput
-                ), f"""File f{input} not exist:
-                    Run .de_interleave() to De-interleave multiplane movies"""
-        '''
-
-        if op.startswith("DI"):
-            print("De-interleaving movies, please wait...")
-            self.deinterleave_output_files = de_interleave(
+        if self.processing_steps.index(op) == 0:  # deinterleave needed
+            de_interleave(
                 focus_files=self.focus_files, efocus=self.efocus
             )
-        if op.startswith("PP"):
-            print("Preprocessing movies, please wait...")
-            preprocess_step(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("BP"):
-            print("Applying bandpass filter, please wait...")
-            spatial_filter_step(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("MC"):
-            print("Applying motion correction, Please wait...")
-            motion_correct_step(
-                translation_files,
-                crop_rect_files,
-                parameters,
-                pairlist,
-                amount_of_files,
-                verbose,
-            )
+
+        if op.startswith("PP") or op.startswith("BP") or op.startswith("MC") or op.startswith("DFF"):
+            pb = progress_bar(amount_of_files,f_message[operation])
+            for input, output in pairlist:
+                f_register[operation](input, output , parameters, verbose)
+                pb.update_progress_bar(1)
         if op.startswith("TR"):
             print("Trim movies, Please wait...")
             trim_movie(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("DFF"):
-            print("Normalizing via DF/F0, Please wait...")
-            create_dff(pairlist, parameters, amount_of_files, overwrite, verbose)
         if op.startswith("PM"):
             print("Projecting Movie, Please wait...")
             project_movie(pairlist, parameters, amount_of_files, overwrite, verbose)
+
         print("done")
 
     @timer
@@ -1356,229 +1305,8 @@ def de_interleave(focus_files: dict, efocus: list, overwrite: bool = False) -> N
         # Update progress bar
         pb.update_progress_bar(1)
         output_files += planes_fs
-    # Need to return value to save the output values back in the object for the remove function later used.
-    return output_files
 
-
-def preprocess_step(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    verbose: bool = False,
-) -> None:
-    """
-    After performing checks, use the isx.preprocess function, which preprocesses movies,
-    optionally applying spatial and temporal downsampling and cropping.
-
-    Parameters
-    ----------
-    pairlist : Tuple[list, list]
-        Tuple containing lists of input and output paths
-    parameters : dict
-       Parameter dictionary of executed functions.
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Preprocessing")
-
-    for input, output in pairlist:
-        if isinstance(parameters["spatial_downsample_factor"], str):
-            res_idx, value = parameters["spatial_downsample_factor"].split("_")
-            if res_idx == "maxHeight":
-                idx_resolution = 0
-            elif res_idx == "maxWidth":
-                idx_resolution = 1
-            else:
-                assert False, "error in sp_downsampling parameter value"
-            movie = isx.Movie.read(input)
-            resolution = movie.spacing.num_pixels
-            del movie
-            parameters["spatial_downsample_factor"] = np.ceil(
-                resolution[idx_resolution] / float(value)
-            )
-
-        parameters.update(
-            {
-                "input_movie_files": os.path.basename(input),
-                "output_movie_files": os.path.basename(output),
-            }
-        )
-
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-
-        isx.preprocess(
-            **parameters_for_isx(
-                parameters,
-                ["comments", "fix_frames_th_std"],
-                {"input_movie_files": input, "output_movie_files": output},
-            )
-        )
-
-        nfixed = fix_frames(output, std_th=parameters["fix_frames_th_std"], report=True)
-
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "preprocess"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} preprocessing completed. {} frames fixed.".format(output, nfixed))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
-
-def spatial_filter_step(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    verbose: bool = False,
-) -> None:
-    """
-    After performing checks, use the isx.spatial_filter function, which
-    apply spatial bandpass filtering to each frame of one or more movies
-
-    Parameters
-    ----------
-    pairlist: Tuple[list, list]
-        Tuple containing lists of input and output paths
-    parameters : dict
-        Parameter dictionary of executed functions.
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Applying Bandpass Filter to")
-
-    for input, output in pairlist:
-        parameters.update(
-            {
-                "input_movie_files": os.path.basename(input),
-                "output_movie_files": os.path.basename(output),
-            }
-        )
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.spatial_filter(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {"input_movie_files": input, "output_movie_files": output},
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "spatial_filter"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} bandpass filtering completed".format(output))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
-
-def motion_correct_step(
-    translation_files: list,
-    crop_rect_files: list,
-    parameters: dict,
-    pairlist: Tuple[list, list],
-    amount_of_files: int,
-    verbose=False,
-) -> None:
-    """
-    After checks, use the isx.motion_correct function, which motion correct movies to a reference frame.
-
-    Parameters
-    ----------
-    translation_files : list
-        A list of file names to write the X and Y translations to. Must be either None,
-        in which case no files are written, or a list of valid file names equal in
-        length to the number of input and output file names
-    crop_rect_files : list
-        The path to a file that will contain the crop rectangle applied to the input
-        movies to generate the output movies
-    parameters : dict
-       Parameter list of executed functions.
-    pairlist : Tuple[list, list]
-        Tuple containing lists of input and output paths
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    """
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Applying Motion Correction to")
-    for i, (input, output) in enumerate(pairlist):
-        new_data = {
-            "input_movie_files": os.path.basename(input),
-            "output_movie_files": os.path.basename(output),
-            "output_translation_files": os.path.basename(translation_files[i]),
-            "output_crop_rect_file": os.path.basename(crop_rect_files[i]),
-        }
-        parameters.update(new_data)
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.motion_correct(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {
-                    "input_movie_files": input,
-                    "output_movie_files": output,
-                    "output_translation_files": translation_files[i],
-                    "output_crop_rect_file": crop_rect_files[i],
-                },
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "motion_correct"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} motion correction completed".format(output))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
+    return
 
 def trim_movie(
     pairlist: Tuple[list, list],
@@ -1793,98 +1521,6 @@ def project_movie(
     print("done")
 
 
-def get_segment_from_movie(
-    inputfile: str,
-    outputfile: str,
-    borders: Iterable,
-    keep_start_time=False,
-    unit="minutes",
-) -> None:
-    """
-    This function gets a segment of a video to create a new one. It's just an easy
-    handler of isx.trim_movie, which trim frames from a movie to produce a new movie
-
-    Parameters
-    ----------
-    inputfile : str
-        Path of the input file to use
-    outputfile : str
-        Path of the output file to use
-    borders: iterable
-        With two elements of the borders in minutes (integers) of the segment.
-        Where negative times means from the end like indexing in a numpy array
-    keep_start_time : Bool, optional
-        If true, keep the start time of the movie, even if some of its
-        initial frames are to be trimmed, by default False
-    unit: str
-        the unit to use for border. It could be 'minutes' or 'seconds'
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    >>> get_segment_from_movie('inputfile.isxd','output.isxd',[-30, -1]) #last 30 minutes
-
-    """
-    assert len(borders) == 2, "borders must have two elements"
-    assert unit in ["minutes", "seconds"], """unit could be 'minutes' or 'seconds'. """
-    movie = isx.Movie.read(inputfile)
-    if unit == "minutes":
-        unitfactor = 60
-    else:
-        unitfactor = 1
-    numframes = movie.timing.num_samples
-    duration_unit = numframes * movie.timing.period.to_msecs() / (1000 * unitfactor)
-    timing_period_unit = movie.timing.period.to_msecs() / (1000 * unitfactor)
-
-    movie.flush()
-
-    crop_segments = []
-    assert (
-        borders[0] > -duration_unit
-    ), "asking for a segment from {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[0] < duration_unit
-    ), "asking for a segment from {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[1] < duration_unit
-    ), "asking for a segment up to {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[1] > -duration_unit
-    ), "asking for a segment up to {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
-
-    # remove fist frames:
-    # don't cut if the segments are from the beggining or just exactlt
-    # duration before the end
-    if borders[0] != 0 and borders[0] != -duration_unit:
-        if borders[0] < 0:
-            end1 = (duration_unit + borders[0]) / timing_period_unit - 1
-        else:
-            end1 = borders[0] / timing_period_unit - 1
-        crop_segments.append([0, int(end1)])
-    # remove last frames:
-    # and don't cut if the segments are up to the end or just the exact
-    # duration
-    if borders[1] != -1 and borders[1] != duration_unit:
-        if borders[1] < 0:
-            start1 = (duration_unit + borders[1] + 1) / timing_period_unit + 1
-        else:
-            start1 = borders[1] / timing_period_unit + 1
-        crop_segments.append([int(start1), numframes])
-
-    if os.path.exists(outputfile):
-        os.remove(outputfile)
-    if len(crop_segments) == 0:
-        print("no trim need it")
-        return
-    isx.trim_movie(
-        input_movie_file=inputfile,
-        output_movie_file=outputfile,
-        crop_segments=np.array(crop_segments),
-        keep_start_time=keep_start_time,
-    )
 
 
 def get_efocus(gpio_file: str) -> list:
