+++ cgk_calcium_tools/isx_aux_functions.py
@@ -1,6 +1,7 @@
 import isx
 import numpy as np
-
+import os
+from typing import Iterable
 
 def create_empty_events(cell_set_file,ed_file):
     """
@@ -67,3 +68,97 @@ def cellset_is_empty(cellset: str, accepted_only:bool = True):
     del cs  # isx keeps the file open otherwise
     
     return is_empty
+
+
+def get_segment_from_movie(
+    inputfile: str,
+    outputfile: str,
+    borders: Iterable,
+    keep_start_time=False,
+    unit="minutes",
+) -> None:
+    """
+    This function gets a segment of a video to create a new one. It's just an easy
+    handler of isx.trim_movie, which trim frames from a movie to produce a new movie
+
+    Parameters
+    ----------
+    inputfile : str
+        Path of the input file to use
+    outputfile : str
+        Path of the output file to use
+    borders: iterable
+        With two elements of the borders in minutes (integers) of the segment.
+        Where negative times means from the end like indexing in a numpy array
+    keep_start_time : Bool, optional
+        If true, keep the start time of the movie, even if some of its
+        initial frames are to be trimmed, by default False
+    unit: str
+        the unit to use for border. It could be 'minutes' or 'seconds'
+
+    Returns
+    -------
+    None
+
+    Examples
+    --------
+    >>> get_segment_from_movie('inputfile.isxd','output.isxd',[-30, -1]) #last 30 minutes
+
+    """
+    assert len(borders) == 2, "borders must have two elements"
+    assert unit in ["minutes", "seconds"], """unit could be 'minutes' or 'seconds'. """
+    movie = isx.Movie.read(inputfile)
+    if unit == "minutes":
+        unitfactor = 60
+    else:
+        unitfactor = 1
+    numframes = movie.timing.num_samples
+    duration_unit = numframes * movie.timing.period.to_msecs() / (1000 * unitfactor)
+    timing_period_unit = movie.timing.period.to_msecs() / (1000 * unitfactor)
+
+    movie.flush()
+
+    crop_segments = []
+    assert (
+        borders[0] > -duration_unit
+    ), "asking for a segment from {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
+    assert (
+        borders[0] < duration_unit
+    ), "asking for a segment from {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
+    assert (
+        borders[1] < duration_unit
+    ), "asking for a segment up to {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
+    assert (
+        borders[1] > -duration_unit
+    ), "asking for a segment up to {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
+
+    # remove fist frames:
+    # don't cut if the segments are from the beggining or just exactlt
+    # duration before the end
+    if borders[0] != 0 and borders[0] != -duration_unit:
+        if borders[0] < 0:
+            end1 = (duration_unit + borders[0]) / timing_period_unit - 1
+        else:
+            end1 = borders[0] / timing_period_unit - 1
+        crop_segments.append([0, int(end1)])
+    # remove last frames:
+    # and don't cut if the segments are up to the end or just the exact
+    # duration
+    if borders[1] != -1 and borders[1] != duration_unit:
+        if borders[1] < 0:
+            start1 = (duration_unit + borders[1] + 1) / timing_period_unit + 1
+        else:
+            start1 = borders[1] / timing_period_unit + 1
+        crop_segments.append([int(start1), numframes])
+
+    if os.path.exists(outputfile):
+        os.remove(outputfile)
+    if len(crop_segments) == 0:
+        print("no trim need it")
+        return
+    isx.trim_movie(
+        input_movie_file=inputfile,
+        output_movie_file=outputfile,
+        crop_segments=np.array(crop_segments),
+        keep_start_time=keep_start_time,
+    )
